09-03-22-02: Modified elevation Data type from Int to Long and loaded data from database.

+++ app/src/main/java/com/example/hikingapp/domain/map/ExtendedMapPoint.kt
@@ -2,7 +2,7 @@ package com.example.hikingapp.domain.map
 
 import com.mapbox.geojson.Point
 
-class ExtendedMapPoint(point: Point, elevation: Int?, var index: Int) : MapPoint(point, elevation) {
+class ExtendedMapPoint(point: Point, elevation: Long?, var index: Int) : MapPoint(point, elevation) {
 
     constructor(point: Point) : this(point, null, -1)
 }
