+++ app/src/main/java/com/example/hikingapp/domain/map/MapPoint.kt
@@ -3,7 +3,7 @@ package com.example.hikingapp.domain.map
 import com.mapbox.geojson.Point
 import java.io.Serializable
 
-open class MapPoint(var point: Point, var elevation: Int?): Serializable{
+open class MapPoint(var point: Point, var elevation: Long?): Serializable{
 
     constructor(point: Point) : this(point,-10000)
 }
